+++ parsers.js
@@ -7,3 +7,4 @@ export { sanitizeString, createHTMLParser, html, htmlToFile } from './parsers/ht
 export { xml } from './parsers/xml.js';
 export { svg } from './parsers/svg.js';
 export { json } from './parsers/json.js';
+export { math } from './parsers/math.js';