+++ parsers/math.js
@@ -1,6 +1,6 @@
-import { elements, attributes } from '@aegisjsproject/sanitizer/config/svg.js';
+import { elements, attributes } from '@aegisjsproject/sanitizer/config/mathml.js';
 
-export function svg(strings, ...args) {
+export function math(strings, ...args) {
 	return Document.parseHTML(
 		String.raw(strings, ...args),
 		{ sanitizer: { elements: elements.concat(['html', 'body', 'head']), attributes }}